restructuring where utilities are placed
