+++ bionetgen/modelapi/bngfile.py
@@ -2,7 +2,7 @@ import bionetgen as bng
 import os, re
 
 from bionetgen.main import BioNetGen
-from .utils import find_BNG_path, run_command, ActionList
+from bionetgen.utils.utils import find_BNG_path, run_command, ActionList
 from tempfile import TemporaryDirectory
 
 # This allows access to the CLIs config setup
