+++ bionetgen/modelapi/bngparser.py
@@ -8,7 +8,7 @@ from .xmlparsers import ParameterBlockXML, CompartmentBlockXML, ObservableBlockX
 from .xmlparsers import SpeciesBlockXML, MoleculeTypeBlockXML, FunctionBlockXML
 from .xmlparsers import RuleBlockXML, EnergyPatternBlockXML, PopulationMapBlockXML
 from .blocks import ActionBlock
-from .utils import ActionList
+from bionetgen.utils.utils import ActionList
 
 # This allows access to the CLIs config setup
 app = BioNetGen()
