+++ bionetgen/modelapi/structs.py
@@ -1,6 +1,6 @@
 from bionetgen.modelapi.pattern import Molecule, Pattern
 from bionetgen.modelapi.rulemod import RuleMod
-from bionetgen.modelapi.utils import ActionList
+from bionetgen.utils.utils import ActionList
 
 
 class ModelObj:
